Only write legacy binary if it doesn't already exist

+++ src/bin/nomic.rs
@@ -404,9 +404,11 @@ fn legacy_bin(config: &nomic::network::Config) -> Result<Option<PathBuf>> {
                 let bin_name = env!("NOMIC_LEGACY_VERSION").trim().replace(" ", "-");
                 let bin_path = bin_dir.join(bin_name);
                 let bin_bytes = include_bytes!(env!("NOMIC_LEGACY_PATH"));
-                log::debug!("Writing legacy binary to {}...", bin_path.display());
-                std::fs::write(&bin_path, bin_bytes).unwrap();
-                std::fs::set_permissions(bin_path, Permissions::from_mode(0o777)).unwrap();
+                if !bin_path.exists() {
+                    log::debug!("Writing legacy binary to {}...", bin_path.display());
+                    std::fs::write(&bin_path, bin_bytes).unwrap();
+                    std::fs::set_permissions(bin_path, Permissions::from_mode(0o777)).unwrap();
+                }
             }
 
             if !bin_dir.exists() {
